+++ src/components/editgrid/EditGrid.js
@@ -271,7 +271,16 @@ export default class EditGridComponent extends NestedArrayComponent {
           flattenedComponents,
           getView: (component, data) => {
             const instance = flattenedComponents[component.key];
-            return instance ? instance.getView(data) : '';
+            let view = instance ? instance.getView(data) : '';
+            if (instance && instance.widget && view !== '--- PROTECTED ---' ) {
+              if (_.isArray(view)) {
+                view = view.map((value) => instance.widget.getValueAsString(value));
+              }
+              else {
+                view = instance.widget.getValueAsString(view);
+              }
+            }
+            return view;
           },
         },
       );